manual-flask-app/requirements.txt
Flask==2.0.2
pymongo==3.12.1
requests==2.26.0
opentelemetry-api==1.17.0
opentelemetry-distro==0.38b0
opentelemetry-instrumentation==0.38b0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-pymongo
opentelemetry-instrumentation-logging